src/main/java/dk/cphbusiness/xpscrumproject/Fag.java
/*
 * To change this license header, choose License Headers in Project Properties.
 * To change this template file, choose Tools | Templates
 * and open the template in the editor.
 */
package dk.cphbusiness.xpscrumproject;

/**
 *
 * @author Muggi
 */
public class Fag implements FagInterface {

    private String title;
    private String description;
    private String teacher;

    public Fag(String title, String description, String teacher) {
        this.title = title;
        this.description = description;
        this.teacher = teacher;
    }
    
    @Override
    public void setTitle(String title) {
        this.title = title;
    }

    @Override
    public String getTitle() {
        return title;
    }

    @Override
    public void setTeacher(String teacher) {
        this.teacher = teacher;
    }

    @Override
    public String getTeacher() {
        return teacher;
    }

    @Override
    public void setDescription(String description) {
        this.description = description;
    }

    @Override
    public String getDescription() {
        return description;
    }

}
